researchflow-production-main/services/worker/requirements-base.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==23.1.0
attrs==25.4.0
Authlib==1.6.6
autograd==1.8.0
autograd-gamma==0.5.0
bcrypt==5.0.0
blinker==1.9.0
boolean.py==5.0
CacheControl==0.14.4
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cyclonedx-python-lib==11.6.0
defusedxml==0.7.1
detect-secrets==1.5.0
dask>=2024.1.0
distlib==0.4.0
dparse==0.6.4
# REMOVED: ecdsa==0.19.1 - CVE-2024-23342 Minerva timing attack, no fix available
# ECDSA operations now use cryptography library via python-jose[cryptography]
email-validator==2.2.0
fastapi==0.128.0
filelock==3.20.3
fonttools==4.61.1
formulaic==1.2.1
gitdb==4.0.12
GitPython==3.1.46
pygit2>=1.13.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx>=0.27.0  # Security: CVE-2024-41991 fix
identify==2.6.15
idna==3.11
iniconfig==2.3.0
interface-meta==1.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
license-expression==30.4.4
lifelines==0.27.8
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.0
matplotlib==3.8.2
mdurl==0.1.2
msgpack==1.1.2
multimethod==1.12
mypy_extensions==1.1.0
narwhals==2.15.0
nltk==3.9.2
nodeenv==1.10.0
numpy==1.26.4
packageurl-python==0.17.6
openpyxl>=3.1.0
packaging==25.0
pandas==2.2.0
pandera==0.18.0
passlib==1.7.4
patsy==1.0.2
pillow==12.1.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.5.1
plotly==6.5.1
pluggy==1.6.0
pre_commit==4.5.1
protobuf>=4.25.0  # Security: JSON recursion depth bypass fix
py-serializable==2.1.0
pyarrow==22.0.0
pyasn1>=0.5.1  # Security: DoS vulnerability fix
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.1
pytest==8.0.0
pytest-asyncio==0.23.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose[cryptography]>=3.3.0  # Security: Algorithm confusion fix (CVE-2024-33664)
python-multipart>=0.0.7  # Security: Arbitrary file write fix (CVE-2024-24762)
pytz==2025.2
PyYAML==6.0.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.19.1
safety==3.7.0
safety-schemas==0.0.16
scikit-learn>=1.4.0  # Security: Data leakage vulnerability fix
scipy==1.11.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.38
starlette>=0.40.0,<0.51.0  # Pinned for fastapi 0.128.0 compatibility
statsmodels==0.14.0
streamlit==1.52.2
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.3.0
tomli_w==1.2.0
tomlkit==0.13.3
tornado==6.5.4
tqdm==4.67.1
typeguard==4.4.4
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
virtualenv==20.36.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.1
reportlab==4.2.0
python-pptx==1.0.2

# Multi-file ingestion dependencies (added 2026-01-27)
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
polars>=0.20.0
duckdb>=0.9.0
xlrd>=2.0.1

# Guideline Engine Dependencies (Stage 20)
redis>=5.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
PyMuPDF>=1.23.0
aiohttp>=3.9.2  # Security: CVE-2024-23334 directory traversal fix

# PHI Detection & Protection (HIPAA compliance)
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0

# Phase 3: App Integration & Monitoring (structlog, Prometheus)
structlog>=24.1.0
prometheus_client>=0.20.0
psutil>=5.9.0
